backend/JobServer/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
build==1.0.3
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.0
fastapi==0.109.0
Flask==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.4
numpy==1.24.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyproject_hooks==1.0.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-jobspy==1.1.35
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
tls-client==0.2.2
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0.post1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
